+++ src/Game/GameObjects/Border.h
@@ -14,7 +14,7 @@ public:
 
     Border(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void render() const override;
-    virtual void update(const uint64_t delta) override;
+    virtual void update(const double delta) override;
 
 private:
     std::shared_ptr<RenderEngine::Sprite> m_sprite;
